implement cli and some refactoring
